backend/requirements.txt
django
djoser
djangorestframework
django-filter
djangorestframework-simplejwt
Pillow
drf-yasg
faker
aiohttp